refactor: moves middleware AuthorizationMiddleware from user to common
